requirements.txt
datasets
transformers
torch
accelerate
sentencepiece
scikit-learn
pandas
numpy
matplotlib
tqdm
pyyaml
seqeval
evaluate
datasketch
rapidfuzz
jsonlines
python-dotenv
groq